Refactor stadnalone-cpal module quite a bit
